refactor: atom 폴더구조 이관
