+++ src/components/templates/cart-select-list/cart-select-list.tsx
@@ -4,7 +4,7 @@ import { css } from '@styled-system/css'
 import { flex, vstack } from '@styled-system/patterns'
 import { IconTrash } from '@tabler/icons-react'
 import { Counter } from '@/components/molecules/counter/counter'
-import { useCartProductSelection, useResetCartProductSelection } from '@/atoms/cart-select-atom'
+import { useCartProductSelection, useResetCartProductSelection } from '@/stores/hooks'
 import { useDeleteCartProductSingle } from '@/mutations/delete-cart-product-single'
 import { useCreateCart } from '@/mutations/create-cart'
 import { useDeleteCartProductAll } from '@/mutations/delete-cart-product-all'
