+++ src/stores/atoms/cart-select-atom.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai'
+
+export const cartSelectAtom = atom(new Set<number>())
